requirements.txt
requests==2.31.0
django==4.2.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
twilio==8.3.0
faker==18.11.1
django-extensions==3.2.3
razorpay==1.3.0
django-environ==0.10.0